type changed to float

+++ api/src/graphql/inputs/productInput.js
@@ -2,7 +2,7 @@ module.exports = `
     input productInput{
         name: String
         description: String,
-        price: Int,
+        price: Float,
         stock: Int,
         image: String,
     }
